Changed the code to be operating system agnostic.
The reason the code was split was due to difference in paths and system
commands. The pathing issue was solved by using the os.path module. The
system commands only occur once, and can be differentiated by checking
the os.name.
